feat: refactor front-end code
