+++ frontend/src/components/error.rs
@@ -1,0 +1,31 @@
+use leptos::prelude::*;
+use std::collections::HashMap;
+
+#[component]
+pub fn ErrorDisplay(
+    erros: ReadSignal<HashMap<String, Vec<String>>>,
+    key: &'static str,
+) -> impl IntoView {
+    view! {
+        {move || {
+            if erros.get().contains_key(key) {
+                let error_messages = erros.get().get(key).cloned().unwrap_or_default();
+                Some(
+                    view! {
+                        <div class="error-message">
+                            <For
+                                each=move || error_messages.clone().into_iter()
+                                key=|msg| msg.clone()
+                                children=move |msg| {
+                                    view! { <p>{msg}</p> }
+                                }
+                            />
+                        </div>
+                    },
+                )
+            } else {
+                None
+            }
+        }}
+    }
+}
