+++ frontend/src/components/machines/mod.rs
@@ -1,0 +1,7 @@
+mod add;
+mod detail;
+mod registered;
+
+pub use add::AddMachine;
+pub use detail::MachineDetailPage;
+pub use registered::RegisteredMachines;
