+++ frontend/src/components/mod.rs
@@ -1,0 +1,9 @@
+mod error;
+mod header;
+mod home;
+mod machines;
+
+pub use error::ErrorDisplay;
+pub use header::Header;
+pub use home::HomePage;
+pub use machines::*;
